test: split preference lifecycle coverage
